// src/app/models/BadgeReport.ts
export class BadgeReport {
  badgeNumber: string;
  cardSeriesNumber: string;
  cardNumber: string;
  dateOfActivation: Date;
  dateOfSecurityCheck: Date;
  dateOfTraining: Date;
  expireDate: Date;
  payment: string;
  deactivated: boolean;
  deactivateReason: string;
  shreddingDate: Date;
  returned: boolean;
  employeeName: string;
  employeeSurname: string;
  occupation: string;
  companyName: string;
  companyNameEn: string;
}
 // { get; set; }
        // public string CardSeriesNumber { get; set; }
        // public string CardNumber { get; set; }
        // public DateTime? DateOfActivation { get; set; }
        // public DateTime? DateOfSecurityCheck { get; set; }
        // public DateTime? DateOfTraining { get; set; }
        // public DateTime? ExpireDate { get; set; }
        // public string Payment { get; set; }
        // public bool Deactivated { get; set; }
        // public string DeactivateReason { get; set; }
        // public DateTime? ShreddingDate { get; set; }
        // public bool Returned { get; set; }
        // public string EmployeeName { get; set; }
        // public string EmployeeSurname { get; set; }
        // public string Occupation { get; set; }
        // public string CompanyName { get; set; }
        // public string CompanyNameEn { get; set; }